move GEDFileBuilder into GEDFile.Builder
